requirements.txt
# StoryMuse - Local AI Co-Author
# Python 3.10+ required

# CLI Framework
typer[all]>=0.9.0

# Terminal UI
rich>=13.0.0

# LLM Client
openai>=1.0.0

# Structured JSON extraction
instructor>=1.0.0

# Data validation
pydantic>=2.0.0

# Environment configuration
python-dotenv>=1.0.0
